Extract MobileCoinClient into an interface.
