+++ .gitignore
@@ -33,6 +33,9 @@ proguard/
 #service key
 service-key.json
 
+# Android Test DS Store
+android-sdk/src/androidTest/res/.DS_Store
+
 # Android Studio Navigation editor temp files
 .navigation/
