+++ android-sdk/src/androidTest/java/com/mobilecoin/lib/TxOutStoreTest.java
@@ -33,7 +33,7 @@ public class TxOutStoreTest {
             throws SerializationException, InvalidFogResponse, NetworkException,
             AttestationException, InvalidUriException {
         AccountKey accountKey = TestKeysManager.getNextAccountKey();
-        MobileCoinClient mobileCoinClient = Environment.makeFreshMobileCoinClient(accountKey);
+        MobileCoinClientImpl mobileCoinClient = Environment.makeFreshMobileCoinClient(accountKey);
         TxOutStore store = new TxOutStore(accountKey);
         store.refresh(
                 mobileCoinClient.viewClient,
